% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecastDFM.R
\name{forecastDFM}
\alias{forecastDFM}
\title{A function to estimate a Dynamic Factor Model and perform forecasts with the estimated model}
\usage{
forecastDFM(X, variable, r, p)
}
\arguments{
\item{X}{A matrix or data frame}

\item{variable}{The variable to be forecast}

\item{r}{The number of factors to be estimated}

\item{p}{The lag length of the ADL model}
}
\value{
The one step ahead forecast of the supplied variable
}
\description{
A function to estimate a Dynamic Factor Model and perform forecasts with the estimated model
}
