% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ICfactors_mine.R
\name{ICfactors_mine}
\alias{ICfactors_mine}
\title{A function to estimate the optimal number of factors for a Dynamic Factor Model.}
\usage{
ICfactors_mine(x, rmax, type = 2)
}
\arguments{
\item{x}{A matrix or data frame}

\item{rmax}{The maximum number of factors to be considered}

\item{type}{Which of the three criterion from Bai and Ng (2002) should be used. (default = 2)}
}
\value{
The optimal number of factors
}
\description{
A function to estimate the optimal number of factors for a Dynamic Factor Model.
}
